do not allow to delete partition on "loop" partition table
libparted uses the virtual "loop" partition table, if it detects
a file system on disk but no partition table. (I.e. the disk has
been directly formatted with a file system). Deleting such a partition
makes libparted unable to detect the partition table (it does not make
up the "loop" partition table anymore), and stops it
from re-adding a partition.

One can still reformat the "partition", or paste into it.

This is at least necessary until gparted
allows formatting a disk with a file system without a partition table.