determine get_selected_device() based on selected_partition
We don't return the device selected in the combo box, but the
device that the currently selected_partition is on. This may
be different e.g. when the filesystem is inside a LUKS container.

Previously removing a partition inside a LUKS container would remove the
partition from the parent device, because GParted was not aware that
the partition table is on a different device than that selected from the
combo box.

But this change allows generally for sub-partitions to reside
on different devices, so it's not at all bound to LUKS. Just think
about complicated hierarchies possible by dm-mapper, lvm, btrfs
snapshots, loop devices etc.

This also makes it possible to show multiple devices (like /dev/sda and
/dev/sdb) together in one tree-view like

/dev/sda
  /dev/sda1 Primary partiton
  /dev/sda2 Extended partition
    /dev/sda3 Logical partition
    /dev/sda4 Logical partition with LUKS
      /dev/mapper/vault1 Primary partition inside LUKS
      /dev/mapper/vault2 Extended partition inside LUKS
        ...
/dev/sdb
  /dev/sdb1
  ...

(The code for actually putting multiple devices in one TreeView is not
included. This code only makes sure that each operation targets the
correct device when multiple devices are shown in one TreeView.)